// src/app/countries/interfaces/country.ts
// Generated by https://quicktype.io

export interface Country {
	name:         Name;
	tld:          string[];
	cca2:         string;
	ccn3:         string;
	cca3:         string;
	cioc:         string;
	independent:  boolean;
	status:       string;
	unMember:     boolean;
	currencies:   Currencies;
	idd:          Idd;
	capital:      string[];
	altSpellings: string[];
	region:       string;
	subregion:    string;
	languages:    Languages;
	translations: { [key: string]: Translation };
	latlng:       number[];
	landlocked:   boolean;
	borders?:     string[];
	area:         number;
	demonyms:     Demonyms;
	flag:         string;
	maps:         Maps;
	population:   number;
	gini?:        { [key: string]: number };
	fifa:         string;
	car:          Car;
	timezones:    string[];
	continents:   string[];
	flags:        Flags;
	coatOfArms:   CoatOfArms;
	startOfWeek:  string;
	capitalInfo:  CapitalInfo;
	postalCode?:  PostalCode;
}

export interface CapitalInfo {
	latlng: number[];
}

export interface Car {
	signs: string[];
	side:  string;
}

export interface CoatOfArms {
	png?: string;
	svg?: string;
}

export interface Currencies {
	USD?: Clp;
	CRC?: Clp;
	DOP?: Clp;
	YER?: Clp;
	CLP?: Clp;
	EUR?: Clp;
}

export interface Clp {
	name:   string;
	symbol: string;
}

export interface Demonyms {
	eng: Eng;
	fra: Eng;
}

export interface Eng {
	f: string;
	m: string;
}

export interface Flags {
	png:  string;
	svg:  string;
	alt?: string;
}

export interface Idd {
	root:     string;
	suffixes: string[];
}

export interface Languages {
	spa?: string;
	eng?: string;
	ara?: string;
	ita?: string;
}

export interface Maps {
	googleMaps:     string;
	openStreetMaps: string;
}

export interface Name {
	common:     string;
	official:   string;
	nativeName: NativeName;
}

export interface NativeName {
	spa?: Translation;
	eng?: Translation;
	ara?: Translation;
	ita?: Translation;
}

export interface Translation {
	official: string;
	common:   string;
}

export interface PostalCode {
	format: string;
	regex:  string;
}